Created members page with image change on hover
Currently, I add a div with the "hidden" class in order to preload the
"funny" images that appear on hover. The better solution would be to
combine the normal and funny images into one image and shift the
background when someone hovers over the picture.
Problem is... I don't have Photoshop!